+++ gcs/CHANGES.md
@@ -2,6 +2,16 @@
 
 ## Next
 
+1. Add support for OAuth2 based client authentication that retrieves
+a refresh token using the authorisation code grant flow configured via the
+following properties:
+
+    ```
+    fs.gs.auth.client.id
+    fs.gs.auth.client.secret
+    fs.gs.auth.refresh.token
+   ```
+
 ## 2.2.9 - 2022-11-09
 
 1.  The Google Cloud Storage Connector now can be used as a